評価のカウントをヘルパー化（投稿評価APIでも使い回すため）
